Put reference above context on home page
It's a small thing, but I think the sections should be put in the order
people need them. If you are coming to the project totally new, you need
to start with some tutorials. When you get up and running, you might
need some how-to guides. Once you have some experience, you just need to
check the reference. Finally, when you want to fill in the gaps you
might want to read more context.