Fix ts includes
Without this config, TypeScript would report this error:

Cannot find module '$app/env' or its corresponding type declarations

This is because `app.d.ts` was not being included in the project. It is
included by `.svelte-kit/tsconfig.json` but by including the tests dir
in `./tsconfig.json` we were *overwriting* the entire include array.