Use pulumilocal
This automatically uses the localstack stack so the user doesn't need to
remember to select it. It also automatically creates the stack if it's
missing.

Before making this change, I was getting an error about "Invalid or
unknown key" on aws.dynamodb.Table. I think it's because
Pulumi.localdev.yaml was missing a region, but I'm not sure. All I know
is that using pulumilocal with the auto-generated Pulumi.localstack.yaml
worked.