Include stackname in resource names
This makes it obvious what resources are for dev/qa etc. when browsing
in the AWS console.

Note that if you capitalize the stackname, it results in an invalid
bucket name. Capital letters are not valid for a bucket name. My guess
is that Pulumi is lower-casing initial characters for us, so
'Dev'->'dev' but 'DEV'-> 'dEV'.