Remove custom domain name from the API.
Certain corporate networks (who will remain nameless) use filtering
software that is too stupid to unblock a backend API that is required
for a site when unblocking the root domain.

We could re-use the same domain for both front and back end, but then
we'd have to disambiguate them with a path and this is just easier.

The only information "exposed" by the raw domain (which users never see)
is that we are deploying to AWS in a specific region, but that
information is in the (publicly available) source code anyway.