Add a delay when voting
This shouldn't be needed, but this test is flaky already. It randomly
fails something like 30% of the time (just hangs waiting for a vote to
register). I'm going to see if this allows it to succeed more reliably.
If it does, then I can try to figure out why a delay is important. It
might be a server side issue that the test has no other way to work
around.